fixed a few bugs. (One of them thanks to a patch from Ricardo). Made more code cleanups.  Introduced a keymap/event system, which will provide a place for changeable keys later. (Derived from a patch Ricardo sent)
SVN-Revision: 102

+++ src/type.c
@@ -21,7 +21,7 @@
 double get_frame_ratio(base_type* pbase)
 {
   unsigned int cur_time = SDL_GetTicks();
-  double frame_ratio = (float)(cur_time-pbase->updated)/(float)(FRAME_RATE);
+  double frame_ratio = ((double)(cur_time-pbase->updated))/((double)FRAME_RATE);
   pbase->updated = cur_time;
   return (frame_ratio == 0 ? 1 : frame_ratio);
 }
